+++ CHANGELOG.md
@@ -4,6 +4,10 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/) 
 and this project adheres to [Semantic Versioning](http://semver.org/).
 
+## [Unreleased]
+### Added
+- Optional debugging mode.
+
 ## 0.0.3 - 2017-02-06
 ### Fixed
 - Use element.nodeName instead of element.constructor.name (fixes problems with phantomjs)